src/requirements.txt
aiolimiter==1.1.0
annotated-types==0.6.0
anyio==4.1.0
blessed==1.20.0
bpython==0.24
certifi==2023.11.17
charset-normalizer==3.3.2
curtsies==0.4.2
cwcwidth==0.1.9
dependency-injector==4.41.0
distro==1.8.0
dnspython==2.4.2
greenlet==3.0.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mongoengine==0.27.0
openai==1.5.0
pip-autoremove==0.10.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
pymongo==4.6.1
python-telegram-bot==20.7
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.12
